[IMP] base: monkeypatch cleanup
Cleans the places where monkeypatching is done outside
of the `monkeypatches` folder.
