+++ odoo/_monkeypatches/re.py
@@ -1,0 +1,6 @@
+import re
+
+
+def patch():
+    """ Default is 512, a little too small for odoo """
+    re._MAXCACHE = 4096
