+++ odoo/cli/server.py
@@ -13,7 +13,6 @@ GNU Public Licence.
 import atexit
 import logging
 import os
-import re
 import sys
 
 from psycopg2.errors import InsufficientPrivilege
@@ -28,7 +27,6 @@ from . import Command
 # Also use the `odoo` logger for the main script.
 _logger = logging.getLogger('odoo')
 
-re._MAXCACHE = 4096  # default is 512, a little too small for odoo
 
 def check_root_user():
     """Warn if the process's user is 'root' (on POSIX system)."""
